Now the characters are finally able to move across the computed path, even if some work on the algorithm is still required.
Some utility methods have been added, like the ones for occupying or
freeing a ResourceProducer. These methods throw two new kinds of
Exception, ControllerAlreadyPresentException and
DifferentOwnerException.

+++ Medieval_POST-APOCALYPSE_logica/src/mpa/core/logic/InterfaceAbstractObject.java
@@ -7,9 +7,9 @@ public interface InterfaceAbstractObject
 
 	public abstract float getY();
 
-	public abstract void setX(int x);
+	public abstract void setX(float x);
 
-	public abstract void setY(int y);
+	public abstract void setY(float y);
 
 	public abstract float getWidth();
